% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio.R
\name{academic_letters}
\alias{academic_letters}
\title{academic_letters}
\usage{
academic_letters(include_area = FALSE)
}
\arguments{
\item{include_area}{logical. If TRUE, includes the area of study in the string and separates each entry by commas.}
}
\description{
Returns string which includes academic letters
}
