% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiliation.R
\name{affiliation}
\alias{affiliation}
\title{affiliation}
\usage{
affiliation(papaja = FALSE)
}
\arguments{
\item{papaja}{if true, affiliations are printed so they can be pasted into a papaja manuscript.}
}
\description{
My affiliation
}
