% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_doi.R
\name{aut_to_doi}
\alias{aut_to_doi}
\title{aut_to_doi}
\usage{
aut_to_doi(string = NULL, n = 3, qry = "")
}
\arguments{
\item{string}{string to search}

\item{n}{results to return}
}
\description{
Find manuscript details based on author year
}
