% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{css}
\alias{css}
\title{css}
\usage{
css(path = NULL, line_height = 1.2, block_margin = 0.075)
}
\arguments{
\item{path}{to file}

\item{line_height}{gap between lines}

\item{block_margin}{margin between blocks}
}
\description{
modify css and save to file
}
