% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pubs.R
\name{export_pubs}
\alias{export_pubs}
\title{export_pubs}
\usage{
export_pubs(path = NULL, format = "bib")
}
\arguments{
\item{path}{path to output file}

\item{format}{format of output file (ris or bib)}
}
\description{
Export publications to ris
}
