% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_doi}
\alias{get_doi}
\title{get_doi}
\usage{
get_doi(title, journal, author)
}
\arguments{
\item{title}{title of article}

\item{journal}{journal of article}

\item{author}{author name}
}
\description{
Get a doi for a publication
}
