% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pubs.R
\name{predict_new_cites}
\alias{predict_new_cites}
\title{predict_new_cites}
\usage{
predict_new_cites(end_date = NULL, id = "m0d4TKcAAAAJ")
}
\arguments{
\item{end_date}{date at which to end at. Defaults to end of the year}
}
\description{
Predict how many cites you'll get in the current year
}
