% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/james_resume_create.R
\name{resume}
\alias{resume}
\title{resume}
\usage{
resume(path = NULL)
}
\arguments{
\item{path}{If provided, the resume is copied to that destination. In this instance, the file is not opened}
}
\description{
open resume
}
