% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/james_resume_create.R
\name{update_alt_resume}
\alias{update_alt_resume}
\title{update_alt_resume}
\usage{
update_alt_resume(
  education = TRUE,
  experience = TRUE,
  professional = TRUE,
  workshops = FALSE,
  supervision = TRUE,
  committees = TRUE,
  packages = TRUE,
  n_pubs = Inf,
  commentaries = TRUE,
  peer_review = TRUE,
  preprints = TRUE,
  conferences = TRUE,
  grants = TRUE,
  path = NULL
)
}
\arguments{
\item{education}{render education}

\item{experience}{render experience}

\item{professional}{render professional experience}

\item{workshops}{include workshops}

\item{supervision}{include supervision}

\item{committees}{include committees}

\item{packages}{include packages}

\item{n_pubs}{numeric, top n cited publications, defaults to Inf}

\item{commentaries}{include commentaries by others}

\item{peer_review}{include peer review activity}

\item{preprints}{include pre-prints}

\item{conferences}{include presentations}

\item{grants}{include grants}

\item{path}{if included, resume is additionally copied to path specified}
}
\value{
The generated HTML path, invisibly.
}
\description{
Create an alternate HTML resume using the canonical resume content.
}
