% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/james_resume_create.R
\name{update_resume}
\alias{update_resume}
\title{update_resume}
\usage{
update_resume(
  push = FALSE,
  education = TRUE,
  experience = TRUE,
  workshops = FALSE,
  supervision = TRUE,
  committees = TRUE,
  packages = TRUE,
  n_pubs = Inf,
  commentaries = TRUE,
  peer_review = TRUE,
  preprints = TRUE,
  conferences = TRUE,
  grants = TRUE,
  path = NULL,
  ...
)
}
\arguments{
\item{push}{if TRUE, changes are pushed to github}

\item{education}{render education}

\item{experience}{render experience}

\item{supervision}{include supervision}

\item{committees}{include committees}

\item{n_pubs}{numeric, top n cited publications, defaults to Inf}

\item{commentaries}{include commentaries by others}

\item{peer_review}{include peer review activity}

\item{conferences}{include conference contributions}

\item{grants}{include grants}

\item{path}{if included, resume is additionally copied to path specified}

\item{...}{additional arguments passed to css}

\item{workshop}{include workshops}

\item{pacakge}{include packages}

\item{preprint}{include pre-prints}
}
\description{
update resume and optionally push changes to github
}
