% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pubs.R
\name{publications}
\alias{publications}
\title{get_pubs}
\usage{
publications(
  id = "m0d4TKcAAAAJ",
  n = Inf,
  journal_exclude = c("ACU Research Bank"),
  cache_doi = TRUE
)
}
\arguments{
\item{id}{google scholar id}

\item{n}{how many pubs to retrieve?}

\item{exclude_journals}{vector of journal names to filter out}
}
\description{
Get and format publications
}
