% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_resume.R
\name{short_resume}
\alias{short_resume}
\title{short_resume}
\usage{
short_resume(path = "short_resume.pdf", n_pubs = 8)
}
\arguments{
\item{path}{path to output file}
}
\description{
Creates a short resume at path
}
